requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
torch==2.2.2
transformers==4.44.2
pillow==10.4.0
requests==2.31.0
chromadb==0.5.3
pytest==8.3.2
nest-asyncio==1.6.0
tenacity==8.3.0
matplotlib==3.8.4